+++ consumer.go
@@ -41,7 +41,7 @@ type OffsetEntry struct {
 	offset rdkafka.Offset
 }
 
-func NewConsumer(id string, topics []string, cfg rdkafka.ConfigMap) *Consumer {
+func NewConsumer(id string, topics []string, cfg rdkafka.ConfigMap) (*Consumer, error) {
 	var err error
 
 	c := Consumer{
@@ -52,12 +52,10 @@ func NewConsumer(id string, topics []string, cfg rdkafka.ConfigMap) *Consumer {
 
 	c.consumer, err = rdkafka.NewConsumer(&cfg)
 	if err != nil {
-		// TODO(agis): make this a log output instead if we ever accept
-		// config from clients
-		c.log.Fatal(err)
+		return nil, err
 	}
 
-	return &c
+	return &c, nil
 }
 
 func (c *Consumer) Run(ctx context.Context) {
@@ -66,7 +64,33 @@ func (c *Consumer) Run(ctx context.Context) {
 	c.log.Printf("Started working...")
 	<-ctx.Done()
 
-	err := c.consumer.Close()
+	// need to drain the consumer queue by calling Poll() until nil is returned
+	// to be sure the following Close() will return
+	// https://github.com/confluentinc/confluent-kafka-go/issues/189#issuecomment-392726037
+	//
+	// Unsubscribe before the Poll() loop, otherwise Poll() will potentially
+	// consume the whole topic
+	//
+	// TODO: remove the workaround once this is fixed
+	err := c.consumer.Unsubscribe()
+	if err != nil {
+		c.log.Printf("Error unsubscribing: %s", err)
+	}
+
+	for {
+		ev := c.consumer.Poll(0)
+		if ev == nil {
+			break
+		} else {
+			if e, ok := ev.(*rdkafka.Message); ok {
+				c.log.Print("Unexpected message when draining consumer:", *e)
+			}
+		}
+	}
+
+	// closing will also trigger a commit if auto commit is enabled
+	// so we don't need to commit explicitly
+	err = c.consumer.Close()
 	if err != nil {
 		c.log.Printf("Error closing: %s", err)
 	}
@@ -82,12 +106,15 @@ func (c *Consumer) Poll(timeoutMS int) (*rdkafka.Message, error) {
 	switch e := ev.(type) {
 	case *rdkafka.Message:
 		return e, nil
+	case rdkafka.OffsetsCommitted:
+		c.log.Print(e)
 	case rdkafka.Error:
 		return nil, errors.New(e.String())
 	default:
 		c.log.Printf("Unknown event type: %T", e)
-		return nil, nil
 	}
+
+	return nil, nil
 }
 
 // SetOffset sets the offset for the given topic and partition to pos.